fix: 수정 삭제 시 컨텐츠 존재 확인

+++ src/main/java/com/cjw/reactivecommunityproject/web/system/resource_management/service/impl/SystemResourceManagementServiceImpl.java
@@ -1,5 +1,6 @@
 package com.cjw.reactivecommunityproject.web.system.resource_management.service.impl;
 
+import com.cjw.reactivecommunityproject.common.exception.model.RcCommonErrorMessage;
 import com.cjw.reactivecommunityproject.common.spring.component.RcUserComponent;
 import com.cjw.reactivecommunityproject.common.spring.model.response.RestResponseVO;
 import com.cjw.reactivecommunityproject.common.spring.pagination.model.request.PaginationRequestVO;
@@ -12,8 +13,8 @@ import com.cjw.reactivecommunityproject.web.system.resource_management.model.ent
 import com.cjw.reactivecommunityproject.web.system.resource_management.model.entity.SystemResourceManagementListEntity;
 import com.cjw.reactivecommunityproject.web.system.resource_management.model.entity.SystemResourceManagementModifyEntity;
 import com.cjw.reactivecommunityproject.web.system.resource_management.model.request.SystemResourceManagementCreateVO;
-import com.cjw.reactivecommunityproject.web.system.resource_management.model.request.SystemResourceManagementModifyVO;
 import com.cjw.reactivecommunityproject.web.system.resource_management.model.request.SystemResourceManagementListVO;
+import com.cjw.reactivecommunityproject.web.system.resource_management.model.request.SystemResourceManagementModifyVO;
 import com.cjw.reactivecommunityproject.web.system.resource_management.service.SystemResourceManagementService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -82,10 +83,16 @@ public class SystemResourceManagementServiceImpl implements SystemResourceManage
 
     @Override
     public RestResponseVO<Void> modify(SystemResourceManagementModifyVO systemResourceManagementModifyVO) {
+        var isExist = systemResourceManagementDao.isExistResourceByUid(systemResourceManagementModifyVO.uid());
+
+        if (Boolean.FALSE.equals(isExist)) {
+            throw new SystemResourceManagementException(SystemResourceManagementErrorMessage.NOT_FOUND_RESOURCE_DETAIL);
+        }
+
         var isOwner = systemResourceManagementDao.isOwner(systemResourceManagementModifyVO.uid(), rcUserComponent.getUserUid());
 
         if (Boolean.FALSE.equals(isOwner)) {
-            throw new SystemResourceManagementException(SystemResourceManagementErrorMessage.UNAUTHORIZED_ACCESS);
+            throw new SystemResourceManagementException(RcCommonErrorMessage.UNAUTHORIZED_ACCESS);
         }
 
         systemResourceManagementDao.updateTransactional(
@@ -107,10 +114,16 @@ public class SystemResourceManagementServiceImpl implements SystemResourceManage
 
     @Override
     public RestResponseVO<Void> remove(Long uid) {
+        var isExist = systemResourceManagementDao.isExistResourceByUid(uid);
+
+        if (Boolean.FALSE.equals(isExist)) {
+            throw new SystemResourceManagementException(SystemResourceManagementErrorMessage.NOT_FOUND_RESOURCE);
+        }
+
         var isOwner = systemResourceManagementDao.isOwner(uid, rcUserComponent.getUserUid());
 
         if (Boolean.FALSE.equals(isOwner)) {
-            throw new SystemResourceManagementException(SystemResourceManagementErrorMessage.UNAUTHORIZED_ACCESS);
+            throw new SystemResourceManagementException(RcCommonErrorMessage.UNAUTHORIZED_ACCESS);
         }
 
         systemResourceManagementDao.deleteTransactional(uid);
